# app/api.yaml
openapi: 3.0.0
info:
  title: Student Report Generation API
  description: API for generating student reports in HTML and PDF formats with AI insights
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /assignment/html/{task_id}:
    get:
      summary: Get HTML report status or content
      description: Check the status of an HTML report or retrieve it if completed
      tags:
        - Reports
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Task ID returned from generation request
      responses:
        200:
          description: Report status or content
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ReportStatus'
                  - $ref: '#/components/schemas/HTMLReportResponse'
        404:
          description: Task not found
...
components:
  schemas:
...
    ReportStatus:
      type: object
      properties:
        status:
          type: string
          enum: [pending, processing, completed, failed]
          description: Current status of the report generation
    HTMLReportResponse:
      type: object
      properties:
        status:
          type: string
          const: completed
        html:
          type: string
          description: HTML content of the report
        student_id:
          type: string
          description: Student ID this report belongs to
...